+++ src/main/java/cart/service/CartService.java
@@ -4,9 +4,9 @@ import cart.dto.CartItemDto;
 import cart.entity.MemberEntity;
 import cart.exception.CartNotFoundException;
 import cart.exception.ProductNotFoundException;
-import cart.repository.CartDao;
-import cart.repository.MemberDao;
-import cart.repository.ProductDao;
+import cart.dao.CartDao;
+import cart.dao.MemberDao;
+import cart.dao.ProductDao;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
