+++ src/test/java/cart/dao/MemberDaoTest.java
@@ -1,4 +1,4 @@
-package cart.repository;
+package cart.dao;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
